Rename the package to db
