+++ sqlmesh/core/state_sync/db/__init__.py
@@ -1,0 +1,3 @@
+from sqlmesh.core.state_sync.db.facade import EngineAdapterStateSync
+
+__all__ = ["EngineAdapterStateSync"]
